+++ .idea/gradle.xml
@@ -12,7 +12,7 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
-            <option value="$PROJECT_DIR$/ekisandroidkeystoresignatureprovider" />
+            <option value="$PROJECT_DIR$/pbtxKeyProvider" />
           </set>
         </option>
         <option name="resolveModulePerSourceSet" value="false" />
