+++ pbtxKeyProvider/src/main/AndroidManifest.xml
