+++ pbtxKeyProvider/src/main/java/ekis/PBTX/PbtxEkis.kt
@@ -2,10 +2,9 @@ package ekis.PBTX
 
 import android.security.keystore.KeyGenParameterSpec
 import android.security.keystore.KeyProperties
+import ekis.PBTX.Model.KeyModel
 import ekis.PBTX.errors.AndroidKeyStoreDeleteError
 import ekis.PBTX.errors.AndroidKeyStoreSigningError
-import com.google.crypto.tink.subtle.EllipticCurves
-import ekis.PBTX.Model.KeyModel
 import ekis.PBTX.errors.ErrorString.Companion.DELETE_KEY_KEYSTORE_GENERIC_ERROR
 import ekis.PBTX.errors.ErrorString.Companion.GENERATE_KEY_ECGEN_MUST_USE_SECP256R1
 import ekis.PBTX.errors.ErrorString.Companion.GENERATE_KEY_KEYGENSPEC_MUST_USE_EC
@@ -13,10 +12,11 @@ import ekis.PBTX.errors.ErrorString.Companion.GENERATE_KEY_MUST_HAS_PURPOSE_SIGN
 import ekis.PBTX.errors.ErrorString.Companion.QUERY_ANDROID_KEYSTORE_GENERIC_ERROR
 import ekis.PBTX.errors.InvalidKeyGenParameter
 import ekis.PBTX.errors.QueryAndroidKeyStoreError
-import java.security.*
-import java.security.interfaces.ECPublicKey
+import ekis.PBTX.utils.PbtxUtils
+import java.security.KeyPairGenerator
+import java.security.KeyStore
+import java.security.Signature
 import java.security.spec.ECGenParameterSpec
-import java.security.spec.X509EncodedKeySpec
 
 /**
  * Utility class provides cryptographic methods to manage keys in the Android KeyStore Signature Provider and uses the keys to sign transactions.
@@ -63,13 +63,8 @@ class PbtxEkis {
             kpg.initialize(keyGenParameterSpec)
             kpg.generateKeyPair()
 
-            var mKey : ByteArray= getProtobufKey(getKeystore(null), alias, null).key
-
-            System.out.println("key:"+mKey)
-
+            var mKey: ByteArray = PbtxUtils.getProtobufMessage(getKeystore(null), alias, null).key
             return mKey
-
-
         }
 
         /**
@@ -78,20 +73,21 @@ class PbtxEkis {
          * The given [alias] is the identity of the key. The new key will be generated with the Default [KeyGenParameterSpec] from the [generateDefaultKeyGenParameterSpecBuilder]
          *
          * @param alias : Alias of the key store
+         * @return Key : byte Array of generated public key
          */
         @JvmStatic
-        fun createKey(alias: String) {
+        fun createKey(alias: String): ByteArray {
             // Create a default KeyGenParameterSpec
             val keyGenParameterSpec: KeyGenParameterSpec =
                     generateDefaultKeyGenParameterSpecBuilder(alias).build()
 
-            generateAndroidKeyStoreKey(keyGenParameterSpec, alias)
+            return generateAndroidKeyStoreKey(keyGenParameterSpec, alias)
         }
 
 
         /**
-         * Get all (SECP256R1) keys in EOS format from Android KeyStore
-         * @return Array of public Keys
+         * Get all (SECP256R1) keys in byte format from Android KeyStore
+         * @return Array of public Keys with the alias.
          */
 
         @Throws(QueryAndroidKeyStoreError::class)
@@ -100,12 +96,10 @@ class PbtxEkis {
             var mList: ArrayList<KeyModel> = ArrayList();
 
             try {
-                val keyStore =
-                        getKeystore(null)
+                val keyStore = getKeystore(null)
                 var aliasList = keyStore.aliases().toList()
                 aliasList.forEach() {
-
-                    var keyModel = getProtobufKey(keyStore, it, null)
+                    var keyModel = PbtxUtils.getProtobufMessage(keyStore, it, null)
                     mList.add(keyModel)
 
                 }
@@ -117,9 +111,8 @@ class PbtxEkis {
             return mList;
         }
 
-
         /**
-         * Delete key in the Android KeyStore with matching the alias name.
+         * Delete key in the Android KeyStore with matching alias name.
          *
          * @param alias : Alias name of the key needs to be deleted.
          */
@@ -136,49 +129,12 @@ class PbtxEkis {
             }
         }
 
-
+        /**
+         * Loading the key store.
+         */
         private fun getKeystore(loadStoreParameter: KeyStore.LoadStoreParameter?): KeyStore {
             return KeyStore.getInstance(ANDROID_KEYSTORE).apply { load(loadStoreParameter) }
-
         }
-
-
-        private fun getProtobufKey(
-                keyStore: KeyStore,
-                alias: String,
-                password: KeyStore.ProtectionParameter?
-        ): KeyModel {
-            val keyEntry = keyStore.getEntry(alias, password) as KeyStore.PrivateKeyEntry
-            val ecPublicKey =
-                    KeyFactory.getInstance(keyEntry.certificate.publicKey.algorithm).generatePublic(
-                            X509EncodedKeySpec(keyEntry.certificate.publicKey.encoded)
-                    ) as ECPublicKey
-
-            val bytes = EllipticCurves.pointEncode(
-                    EllipticCurves.CurveType.NIST_P256,
-                    EllipticCurves.PointFormatType.COMPRESSED,
-                    ecPublicKey.w
-            )
-
-            val i = 1
-            val b = i.toByte()
-            val ba = ByteArray(1)
-            ba[0] = b
-            // create a destination array that is the size of the two arrays
-            val destination = ByteArray(bytes.size + ba.size)
-
-            System.arraycopy(ba, 0, destination, 0, 1)
-            System.arraycopy(bytes, 0, destination, 1, bytes.size)
-
-            return protobufKeyModel(destination, alias)
-        }
-
-        fun ByteArray.toHexString(): String {
-            return this.joinToString("") {
-                java.lang.String.format("%02x", it)
-            }
-        }
-
 
         /**
          * Sign data with a key in the KeyStore.
@@ -190,10 +146,7 @@ class PbtxEkis {
          */
         @Throws(AndroidKeyStoreSigningError::class)
         @JvmStatic
-        fun signData(
-                data: ByteArray,
-                alias: String
-        ): ByteArray? {
+        fun signData(data: ByteArray, alias: String): ByteArray? {
             try {
                 var keyStore = getKeystore(null)
                 val keyEntry = keyStore.getEntry(alias, null) as KeyStore.PrivateKeyEntry
@@ -208,17 +161,6 @@ class PbtxEkis {
                 throw AndroidKeyStoreSigningError(ex)
             }
         }
-
-        private fun protobufKeyModel(key: ByteArray, alias: String): KeyModel {
-
-            var keyModel = KeyModel();
-            keyModel.alias = alias
-            keyModel.key = key
-
-            return keyModel
-
-        }
-
 
         /**
          * Generate a default [KeyGenParameterSpec.Builder] with
